sessions: More structuring for routes.
