--- app/features/engine/global.ts
@@ -1,1 +1,0 @@
-export const sessionData: any = {};
