+++ app/features/index.ts
@@ -34,7 +34,7 @@ import "./common/timeOutSession";
 import "./ipv/idx";
 import "./ipv/info";
 import "./ipv/json";
-import "./ipv/oauth2/token";
-import "./ipv/oauth2/authorize";
-import "./ipv/oauth2/userinfo";
+import "./oauth2/token";
+import "./oauth2/authorize";
+import "./oauth2/userinfo";
 import "./passport/start";
