--- app/features/ipv/oauth2/token/index.ts
@@ -1,3 +1,0 @@
-import { postOAuthToken } from "./controller";
-
-export { postOAuthToken };
