+++ app/features/oauth2/authorize/index.ts
@@ -1,0 +1,3 @@
+import { SetupOAuthAuthorizeController, getAuthorize } from "./controller";
+
+export { SetupOAuthAuthorizeController, getAuthorize };
