+++ app/features/oauth2/token/index.ts
@@ -1,0 +1,3 @@
+import { SetupOAuthTokenController, postOAuthToken } from "./controller";
+
+export { SetupOAuthTokenController, postOAuthToken };
